refactor: (#845) 모범학생 조회 DTO를 VoteDTO로 이동
