+++ dms-core/src/main/kotlin/team/aliens/dms/domain/student/service/GetStudentService.kt
@@ -8,7 +8,7 @@ import team.aliens.dms.domain.point.spi.vo.StudentWithPointVO
 import team.aliens.dms.domain.room.model.Room
 import team.aliens.dms.domain.student.model.Student
 import team.aliens.dms.domain.student.spi.vo.AllStudentsVO
-import team.aliens.dms.domain.vote.dto.response.ModelStudentResponse
+import team.aliens.dms.domain.vote.dto.reponse.ModelStudentResponse
 import java.time.LocalDate
 import java.util.UUID
