+++ dms-core/src/main/kotlin/team/aliens/dms/domain/student/usecase/GetModelStudentsUseCase.kt
@@ -2,7 +2,7 @@ package team.aliens.dms.domain.student.usecase
 
 import team.aliens.dms.common.annotation.ReadOnlyUseCase
 import team.aliens.dms.domain.student.service.StudentService
-import team.aliens.dms.domain.vote.dto.response.ModelStudentResponse
+import team.aliens.dms.domain.vote.dto.reponse.ModelStudentResponse
 import java.time.LocalDate
 
 @ReadOnlyUseCase
