--- dms-core/src/main/kotlin/team/aliens/dms/domain/vote/dto/response/ModelStudentsResponse.kt
@@ -1,6 +1,0 @@
-package team.aliens.dms.domain.vote.dto.response
-
-data class ModelStudentsResponse(
-    val students: List<ModelStudentResponse>
-
-)
